requirements.txt
json
os
numpy
argparse
whois
socket
concurrent.futures
dns.resolver
requests
urllib
re
io
time
csv
pymongo
tensorflow
torch
array
pickle
dotenv
logging
dns.resolver
OpenSSL
datetime
git+https://github.com/rthalley/dnspython